requirements.txt
pydantic>=2.9.0
python-dotenv>=1.0.1
langchain>=0.3.0
langgraph>=0.2.30
chromadb>=0.5.5
langchain-chroma
langchain-community>=0.2.10
langchain-huggingface>=0.0.3
sentence-transformers>=3.0.1
pypdf>=4.3.1
beautifulsoup4>=4.12.3
unstructured>=0.14.10
langchain-openai>=0.2.0
rank-bm25
duckduckgo-search>=5.0.0
# LlamaIndex dependencies
llama-index>=0.10.0
llama-index-vector-stores-chroma
llama-index-embeddings-huggingface
llama-index-retrievers-bm25
# OCR dependencies
pytesseract
pdf2image
pymupdf

EbookLib
html2text

